Move `LargeGridIcon` to Icons
